ui: broken down the menu components
