+++ web/core/components/workspace-notifications/sidebar/filters/index.ts
@@ -1,2 +1,2 @@
-export * from "./root";
+export * from "./menu";
 export * from "./applied-filter";
