+++ web/core/components/workspace-notifications/sidebar/filters/menu/index.ts
@@ -1,0 +1,2 @@
+export * from "./root";
+export * from "./menu-option-item";
