+++ web/core/components/workspace-notifications/sidebar/header/options/menu-option/index.ts
@@ -1,0 +1,2 @@
+export * from "./root";
+export * from "./menu-item";
